API/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.2
tensorflow-cpu==2.15.0
keras==2.15.0
nltk==3.8.1
numpy==1.26.2
gensim==4.3.2
scipy==1.11.4